requirements.txt
seaborn
pandas
python-dotenv
numpy
requests
openai
